Add hascolumn before drop column email in authors

+++ database/migrations/2024_06_13_123436_add_email_to_authors.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('authors', function (Blueprint $table) {
-            $table->dropColumn('email');
-        });
+        if(Schema::hasColumn('authors', 'email')) {
+            Schema::table('authors', function (Blueprint $table) {
+                $table->dropColumn('email');
+            });
+        }
     }
 };
